ref(trace-explorer): Split traces hooks into separate files
Refactoring the trace explorer into smaller pieces. Starting with splitting the
hooks out into a separate folder.
